fix: github api fallback and refactor setup
