+++ src/commands/setup.ts
@@ -2,102 +2,39 @@ import inquirer from "inquirer";
 import fs from "fs-extra";
 import path from "path";
 import os from "os";
-import { execSync } from "child_process";
 import chalk from "chalk";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
+import { setupShellShortcuts } from "./setup/shell-shortcuts.js";
+import { setupCodexSymlink, setupOpenCodeSymlink } from "./setup/symlinks.js";
+import { checkAndInstallDependencies } from "./setup/dependencies.js";
+import { updateSettings, type SetupOptions } from "./setup/settings.js";
+import {
+  SimpleSpinner,
+  downloadFromGitHub,
+  downloadDirectoryFromGitHub,
+} from "./setup/utils.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
 
-// Simple spinner replacement for clack.spinner()
-class SimpleSpinner {
-  private message: string = "";
-
-  start(message: string) {
-    this.message = message;
-    console.log(chalk.gray(`⏳ ${message}...`));
-  }
-
-  stop(message: string) {
-    console.log(chalk.green(`✓ ${message}`));
-  }
-}
-
-interface SetupOptions {
-  shellShortcuts: boolean;
-  commandValidation: boolean;
-  customStatusline: boolean;
-  aiblueprintCommands: boolean;
-  aiblueprintAgents: boolean;
-  outputStyles: boolean;
-  notificationSounds: boolean;
-  codexSymlink: boolean;
-}
-
 const GITHUB_RAW_BASE =
   "https://raw.githubusercontent.com/Melvynx/aiblueprint-cli/main/claude-code-config";
 
-async function downloadFromGitHub(
-  relativePath: string,
-  targetPath: string,
-): Promise<boolean> {
-  try {
-    const url = `${GITHUB_RAW_BASE}/${relativePath}`;
-    const response = await fetch(url);
-    if (!response.ok) {
-      return false;
-    }
-    const content = await response.arrayBuffer();
-    await fs.ensureDir(path.dirname(targetPath));
-    await fs.writeFile(targetPath, Buffer.from(content));
-    return true;
-  } catch (error) {
-    return false;
-  }
+export interface SetupCommandParams {
+  claudeCodeFolder?: string;
+  codexFolder?: string;
+  openCodeFolder?: string;
+  skipInteractive?: boolean;
 }
 
-async function downloadDirectoryFromGitHub(
-  dirPath: string,
-  targetDir: string,
-): Promise<boolean> {
-  try {
-    // Use GitHub API to list directory contents
-    const apiUrl = `https://api.github.com/repos/Melvynx/aiblueprint-cli/contents/claude-code-config/${dirPath}`;
-    const response = await fetch(apiUrl);
-    if (!response.ok) {
-      return false;
-    }
-
-    const files = await response.json();
-    if (!Array.isArray(files)) {
-      return false;
-    }
-
-    await fs.ensureDir(targetDir);
-
-    for (const file of files) {
-      const relativePath = `${dirPath}/${file.name}`;
-      const targetPath = path.join(targetDir, file.name);
-
-      if (file.type === "file") {
-        await downloadFromGitHub(relativePath, targetPath);
-      } else if (file.type === "dir") {
-        await downloadDirectoryFromGitHub(relativePath, targetPath);
-      }
-    }
-
-    return true;
-  } catch (error) {
-    return false;
-  }
-}
-
-export async function setupCommand(
-  customClaudeCodeFolder?: string,
-  customCodexFolder?: string,
-  skipInteractive?: boolean,
-) {
+export async function setupCommand(params: SetupCommandParams = {}) {
+  const {
+    claudeCodeFolder: customClaudeCodeFolder,
+    codexFolder: customCodexFolder,
+    openCodeFolder: customOpenCodeFolder,
+    skipInteractive,
+  } = params;
   try {
     console.log(chalk.blue.bold("\n🚀 AIBlueprint Claude Code Setup\n"));
     console.log(chalk.bgBlue(" Setting up your Claude Code environment "));
@@ -114,6 +51,7 @@ export async function setupCommand(
         "outputStyles",
         "notificationSounds",
         "codexSymlink",
+        "openCodeSymlink",
       ];
       console.log(chalk.green("✓ Installing all features (--skip mode)"));
     } else {
@@ -161,7 +99,12 @@ export async function setupCommand(
             {
               value: "codexSymlink",
               name: "Codex symlink - Link commands to ~/.codex/prompts",
-              checked: true,
+              checked: false,
+            },
+            {
+              value: "openCodeSymlink",
+              name: "OpenCode symlink - Link commands to ~/.config/opencode/command",
+              checked: false,
             },
           ],
         },
@@ -184,6 +127,7 @@ export async function setupCommand(
       outputStyles: features.includes("outputStyles"),
       notificationSounds: features.includes("notificationSounds"),
       codexSymlink: features.includes("codexSymlink"),
+      openCodeSymlink: features.includes("openCodeSymlink"),
     };
 
     const s = new SimpleSpinner();
@@ -298,8 +242,22 @@ export async function setupCommand(
 
     if (options.codexSymlink && options.aiblueprintCommands) {
       s.start("Setting up Codex symlink");
-      await setupCodexSymlink(claudeDir, customCodexFolder, customClaudeCodeFolder);
+      await setupCodexSymlink(
+        claudeDir,
+        customCodexFolder,
+        customClaudeCodeFolder,
+      );
       s.stop("Codex symlink configured");
+    }
+
+    if (options.openCodeSymlink && options.aiblueprintCommands) {
+      s.start("Setting up OpenCode symlink");
+      await setupOpenCodeSymlink(
+        claudeDir,
+        customOpenCodeFolder,
+        customClaudeCodeFolder,
+      );
+      s.stop("OpenCode symlink configured");
     }
 
     if (options.aiblueprintAgents) {
@@ -397,239 +355,4 @@ export async function setupCommand(
     console.log(chalk.red("Setup failed!"));
     process.exit(1);
   }
-}
-
-async function setupShellShortcuts() {
-  try {
-    const platform = os.platform();
-    let shellConfigFile: string;
-
-    if (platform === "darwin") {
-      shellConfigFile = path.join(os.homedir(), ".zshenv");
-    } else if (platform === "linux") {
-      const shell = process.env.SHELL || "";
-      if (shell.includes("zsh")) {
-        shellConfigFile = path.join(os.homedir(), ".zshrc");
-      } else {
-        shellConfigFile = path.join(os.homedir(), ".bashrc");
-      }
-    } else {
-      console.log(
-        chalk.yellow("Shell shortcuts are only supported on macOS and Linux"),
-      );
-      return;
-    }
-
-    const aliases = `
-# AIBlueprint Claude Code aliases
-alias cc="claude --dangerously-skip-permissions"
-alias ccc="claude --dangerously-skip-permissions -c"
-`;
-
-    const existingContent = await fs
-      .readFile(shellConfigFile, "utf-8")
-      .catch(() => "");
-
-    if (!existingContent.includes("AIBlueprint Claude Code aliases")) {
-      await fs.appendFile(shellConfigFile, aliases);
-    }
-  } catch (error) {
-    console.error(chalk.red("Error setting up shell shortcuts:"), error);
-    throw error;
-  }
-}
-
-async function setupCodexSymlink(
-  claudeDir: string,
-  customCodexFolder?: string,
-  customClaudeCodeFolder?: string,
-) {
-  try {
-    let codexDir: string;
-    if (customCodexFolder) {
-      codexDir = path.resolve(customCodexFolder);
-    } else if (customClaudeCodeFolder) {
-      const parentDir = path.dirname(claudeDir);
-      codexDir = path.join(parentDir, "codex");
-    } else {
-      codexDir = path.join(os.homedir(), ".codex");
-    }
-    const promptsPath = path.join(codexDir, "prompts");
-    const commandsPath = path.join(claudeDir, "commands");
-
-    await fs.ensureDir(codexDir);
-
-    const promptsExists = await fs.pathExists(promptsPath);
-    if (promptsExists) {
-      const stat = await fs.lstat(promptsPath);
-      if (stat.isSymbolicLink()) {
-        await fs.remove(promptsPath);
-      } else {
-        console.log(
-          chalk.yellow(
-            "  ~/.codex/prompts already exists and is not a symlink. Skipping...",
-          ),
-        );
-        return;
-      }
-    }
-
-    await fs.symlink(commandsPath, promptsPath);
-  } catch (error) {
-    console.error(chalk.red("Error setting up Codex symlink:"), error);
-    throw error;
-  }
-}
-
-async function checkAndInstallDependencies() {
-  const checkCommand = (cmd: string): boolean => {
-    try {
-      execSync(`which ${cmd}`, { stdio: "ignore" });
-      return true;
-    } catch {
-      return false;
-    }
-  };
-
-  if (!checkCommand("bun")) {
-    console.log(chalk.yellow("\n  Installing bun..."));
-    try {
-      execSync("npm install -g bun", { stdio: "inherit" });
-    } catch (error) {
-      console.log(
-        chalk.red(
-          "  Failed to install bun. Please install it manually: npm install -g bun",
-        ),
-      );
-    }
-  }
-
-  if (!checkCommand("ccusage")) {
-    console.log(chalk.yellow("\n  Installing ccusage..."));
-    try {
-      execSync("npm install -g ccusage", { stdio: "inherit" });
-    } catch (error) {
-      console.log(
-        chalk.red(
-          "  Failed to install ccusage. Please install it manually: npm install -g ccusage",
-        ),
-      );
-    }
-  }
-}
-
-async function updateSettings(options: SetupOptions, claudeDir: string) {
-  const settingsPath = path.join(claudeDir, "settings.json");
-  let settings: any = {};
-
-  try {
-    const existingSettings = await fs.readFile(settingsPath, "utf-8");
-    settings = JSON.parse(existingSettings);
-  } catch {
-    // Settings file doesn't exist or is invalid
-  }
-
-  if (options.customStatusline) {
-    if (settings.statusLine) {
-      const confirmAnswer = await inquirer.prompt([
-        {
-          type: "confirm",
-          name: "replace",
-          message: "You already have a statusLine configuration. Replace it?",
-        },
-      ]);
-
-      if (!confirmAnswer.replace) {
-        console.log(
-          chalk.yellow("  Keeping existing statusLine configuration"),
-        );
-      } else {
-        settings.statusLine = {
-          type: "command",
-          command: `bash ${path.join(claudeDir, "scripts/statusline-ccusage.sh")}`,
-          padding: 0,
-        };
-      }
-    } else {
-      settings.statusLine = {
-        type: "command",
-        command: `bash ${path.join(claudeDir, "scripts/statusline-ccusage.sh")}`,
-        padding: 0,
-      };
-    }
-  }
-
-  if (!settings.hooks) {
-    settings.hooks = {};
-  }
-
-  if (options.commandValidation) {
-    if (!settings.hooks.PreToolUse) {
-      settings.hooks.PreToolUse = [];
-    }
-
-    const bashHook = {
-      matcher: "Bash",
-      hooks: [
-        {
-          type: "command",
-          command: `bun ${path.join(claudeDir, "scripts/validate-command.js")}`,
-        },
-      ],
-    };
-
-    const existingBashHook = settings.hooks.PreToolUse.find(
-      (h: any) => h.matcher === "Bash",
-    );
-    if (!existingBashHook) {
-      settings.hooks.PreToolUse.push(bashHook);
-    }
-  }
-
-  if (options.notificationSounds) {
-    if (!settings.hooks.Stop) {
-      settings.hooks.Stop = [];
-    }
-
-    const stopHook = {
-      matcher: "",
-      hooks: [
-        {
-          type: "command",
-          command: `afplay -v 0.1 ${path.join(claudeDir, "song/finish.mp3")}`,
-        },
-      ],
-    };
-
-    const existingStopHook = settings.hooks.Stop.find((h: any) =>
-      h.hooks?.some((hook: any) => hook.command?.includes("finish.mp3")),
-    );
-    if (!existingStopHook) {
-      settings.hooks.Stop.push(stopHook);
-    }
-
-    if (!settings.hooks.Notification) {
-      settings.hooks.Notification = [];
-    }
-
-    const notificationHook = {
-      matcher: "",
-      hooks: [
-        {
-          type: "command",
-          command: `afplay -v 0.1 ${path.join(claudeDir, "song/need-human.mp3")}`,
-        },
-      ],
-    };
-
-    const existingNotificationHook = settings.hooks.Notification.find(
-      (h: any) =>
-        h.hooks?.some((hook: any) => hook.command?.includes("need-human.mp3")),
-    );
-    if (!existingNotificationHook) {
-      settings.hooks.Notification.push(notificationHook);
-    }
-  }
-
-  await fs.writeJson(settingsPath, settings, { spaces: 2 });
 }
